Extract `EpicAssets` class into its own class file and delete unused `PieceDef` class.
Notes:
* `PieceDef` was added 4 years ago in commit d3d1722 and its usage was removed in the Jotunn conversion in commit 5cebd74.
