+++ EpicLoot/EpicLoot.cs
@@ -26,48 +26,6 @@ using Object = UnityEngine.Object;
 
 namespace EpicLoot
 {
-    public sealed class EpicAssets
-    {
-        public AssetBundle AssetBundle;
-        public static Dictionary<string, Object> AssetCache = new Dictionary<string, Object>();
-
-        public Sprite EquippedSprite;
-        public Sprite AugaEquippedSprite;
-        public Sprite GenericSetItemSprite;
-        public Sprite AugaSetItemSprite;
-        public Sprite GenericItemBgSprite;
-        public Sprite AugaItemBgSprite;
-        public Sprite EnchantmentSparkle;
-        public GameObject[] MagicItemLootBeamPrefabs = new GameObject[5];
-        public readonly Dictionary<string, GameObject[]> CraftingMaterialPrefabs = new Dictionary<string, GameObject[]>();
-        public Sprite SmallButtonEnchantOverlay;
-        public Sprite DodgeBuffSprite;
-        public AudioClip[] MagicItemDropSFX = new AudioClip[5];
-        public AudioClip ItemLoopSFX;
-        public AudioClip AugmentItemSFX;
-        public GameObject MerchantPanel;
-        public Sprite MapIconTreasureMap;
-        public Sprite MapIconBounty;
-        public AudioClip AbandonBountySFX;
-        public AudioClip DoubleJumpSFX;
-        public AudioClip DodgeBuffSFX;
-        public AudioClip OffSetSFX;
-        public GameObject DebugTextPrefab;
-        public GameObject AbilityBar;
-        public GameObject WelcomMessagePrefab;
-
-        public const string DummyName = "EL_DummyPrefab";
-        public static GameObject DummyPrefab() => PrefabManager.Instance.GetPrefab(DummyName);
-    }
-
-    public sealed class PieceDef
-    {
-        public string Table;
-        public string CraftingStation;
-        public string ExtendStation;
-        public List<RecipeRequirementConfig> Resources = new List<RecipeRequirementConfig>();
-    }
-
     [BepInPlugin(PluginId, DisplayName, Version)]
     [BepInDependency(Jotunn.Main.ModGuid)]
     [BepInDependency("com.ValheimModding.NewtonsoftJsonDetector")]
